requirements.txt
torch>=2.6.0
nltk>=3.9.1
jieba>=0.42.1
sentencepiece>=0.2.0
matplotlib>=3.10.0
seaborn>=0.13.2
pyyaml>=5.4.0